% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{AddChild}
\alias{AddChild}
\title{Append a child element to a parent element}
\usage{
AddChild(element, child)
}
\arguments{
\item{element}{parent element}

\item{child}{element to be appended to the parent element}
}
\value{
parent element with the appended child
}
\description{
Append a child element to a parent element
}
