% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{AddDAG}
\alias{AddDAG}
\title{Add a sub-DAG (synonym for addJob)}
\usage{
AddDAG(workflow, dag)
}
\arguments{
\item{workflow}{Workflow object}

\item{dag}{Sub-DAG to be appended}
}
\value{
The Workflow object with the sub-DAG appended
}
\description{
Add a sub-DAG (synonym for addJob)
}
\seealso{
\code{\link{Workflow}}, \code{\link{AddJob}}, \code{\link{DAG}}
}
