% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{AddDAX}
\alias{AddDAX}
\title{Add a sub-DAX (synonym for addJob)}
\usage{
AddDAX(workflow, dax)
}
\arguments{
\item{workflow}{Workflow object}

\item{dax}{Sub-DAX to be appended}
}
\value{
The Workflow object with the sub-DAX appended
}
\description{
Add a sub-DAX (synonym for addJob)
}
\seealso{
\code{\link{Workflow}}, \code{\link{AddJob}}, \code{\link{DAX}}
}
