% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{AddDependency}
\alias{AddDependency}
\title{Add a dependency to the workflow}
\usage{
AddDependency(workflow, dep)
}
\arguments{
\item{workflow}{The Workflow object}

\item{dep}{The dependency object}
}
\value{
The Workflow object containing the dependency
}
\description{
Add a dependency to the workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Dependency}},
 \code{\link{Depends}}, \code{\link{RemoveDependency}}
}
