% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{AddExecutable}
\alias{AddExecutable}
\title{Add an executable to the Workflow}
\usage{
AddExecutable(workflow, executable)
}
\arguments{
\item{workflow}{Workflow object}

\item{executable}{Executable object}
}
\value{
The Workflow object with the executable appended
}
\description{
Add an executable to the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Executable}},
 \code{\link{RemoveExecutable}}, \code{\link{ClearExecutables}}
}
