% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{AddFile}
\alias{AddFile}
\title{Add a file to the Workflow}
\usage{
AddFile(workflow, file)
}
\arguments{
\item{workflow}{Workflow object}

\item{file}{File object}
}
\value{
The Workflow object with the file appended
}
\description{
Add a file to the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{File}}, \code{\link{RemoveFile}}, \code{\link{ClearFiles}}
}
