% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{AddJob}
\alias{AddJob}
\title{Add a job to the Workflow}
\usage{
AddJob(workflow, job)
}
\arguments{
\item{workflow}{Workflow object}

\item{job}{Job object}
}
\value{
Workflow object with the job appended
}
\description{
Add a job to the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{RemoveJob}}
}
