% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executable.R, R/file.R, R/generic_functions.R
\name{AddPFN.Executable}
\alias{AddPFN.Executable}
\alias{AddPFN.File}
\alias{AddPFN}
\title{Add a PFN to the object}
\usage{
\method{AddPFN}{Executable}(obj, pfn)

\method{AddPFN}{File}(obj, pfn)

AddPFN(obj, pfn)
}
\arguments{
\item{obj}{Object to append the PFN}

\item{pfn}{The PFN}
}
\value{
The object containing the PFN
}
\description{
Add a PFN to the object
}
\seealso{
\code{\link{Executable}}

\code{\link{File}}

\code{\link{PFN}}
}
