% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R, R/dax.R, R/executable.R, R/file.R,
%   R/generic_functions.R, R/job.R, R/pfn.R
\name{AddProfile.DAG}
\alias{AddProfile.DAG}
\alias{AddProfile.DAX}
\alias{AddProfile.Executable}
\alias{AddProfile.File}
\alias{AddProfile}
\alias{AddProfile.Job}
\alias{AddProfile.PFN}
\title{Add a profile to the object}
\usage{
\method{AddProfile}{DAG}(obj, profile)

\method{AddProfile}{DAX}(obj, profile)

\method{AddProfile}{Executable}(obj, profile)

\method{AddProfile}{File}(obj, profile)

AddProfile(obj, profile)

\method{AddProfile}{Job}(obj, profile)

\method{AddProfile}{PFN}(obj, profile)
}
\arguments{
\item{obj}{Object to append the profile}

\item{profile}{The profile}
}
\value{
The object containing the profile
}
\description{
Add a profile to the object
}
\seealso{
\code{\link{DAG}}

\code{\link{DAX}}

\code{\link{Executable}}

\code{\link{File}}

\code{\link{Profile}}

\code{\link{Job}}

\code{\link{PFN}}
}
