% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_mixin.R
\name{AddProfileMixin}
\alias{AddProfileMixin}
\title{Add a profile to the object}
\usage{
AddProfileMixin(profile.mixin, profile)
}
\arguments{
\item{profile.mixin}{Profile_mixin object}

\item{profile}{Profile object to be added}
}
\value{
The profile_mixin object with the profile appended
}
\description{
Add a profile to the object
}
\seealso{
\code{\link{Profile}}
}
