% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{AddTransformation}
\alias{AddTransformation}
\title{Add a transformation to the Workflow}
\usage{
AddTransformation(workflow, transformation)
}
\arguments{
\item{workflow}{Workflow object}

\item{transformation}{Transformation object}
}
\value{
The Workflow object with the transformation appended
}
\description{
Add a transformation to the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Transformation}},
 \code{\link{RemoveTransformation}}, \code{\link{ClearTransformations}}
}
