% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_directories.R
\name{assign_directories}
\alias{assign_directories}
\title{Title}
\usage{
assign_directories(base = "")
}
\arguments{
\item{base}{The root directory of your project}
}
\value{
DIR, a list containing the filenames of your directories
}
\description{
Title
}
\examples{
\dontrun{
DIR <- assign_directories(base = "your working directory")
}
}
