% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{assure_package_installation}
\alias{assure_package_installation}
\title{Maks sure the list of packages is installed}
\usage{
assure_package_installation(
  path = default_pkglist(),
  lib = assure_user_lib()[1],
  repos = assure_repos(),
  dependencies = c("Depends", "Imports")
)
}
\arguments{
\item{path}{path to list of packages (one per line)}
}
\description{
Maks sure the list of packages is installed
}
