% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{assure_repos}
\alias{assure_repos}
\title{Make sure the repos option is set and matches the environmental variable}
\usage{
assure_repos()
}
\value{
URL for repos to use
}
\description{
Make sure the repos option is set and matches the environmental variable
}
