% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{assure_user_lib}
\alias{assure_user_lib}
\title{Make sure the path to the user's location for R packages exists}
\usage{
assure_user_lib(lib_path = Sys.getenv("R_LIBS_USER"))
}
\arguments{
\item{lib_path}{where to put packages}
}
\value{
detected library paths
}
\description{
Make sure the path to the user's location for R packages exists
}
