% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{available_dependencies}
\alias{available_dependencies}
\title{List available packages}
\usage{
available_dependencies(packages, repos = assure_repos(), ...)
}
\arguments{
\item{packages}{packages to check dependency info for}

\item{...}{additional options to \code{available.packages}}
}
\value{
list of lists of packages with dependencies and versions
}
\description{
List available packages
}
