% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environ.R
\name{build_environ_list}
\alias{build_environ_list}
\title{Create a list of values to send to .Renviron file}
\usage{
build_environ_list(
  ...,
  .env = rlang::new_environment(parent = rlang::caller_env())
)
}
\arguments{
\item{...}{values to modify from defaults}

\item{.env}{environment to evaluate the arguments in,
child of caller by default}
}
\value{
list of environment values
}
\description{
Any default-required values missing from inputs are filled
in from defaults (see \code{?get_default_environ})
}
