% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-geocoder.R
\name{census_download_us_state_cousub_file}
\alias{census_download_us_state_cousub_file}
\title{Retrieve the COUSUB file for a given place/time}
\usage{
census_download_us_state_cousub_file(
  state,
  year = "2020",
  cache_dir = workflow::build_dir("census-geocoder-addrfeat-cache")
)
}
\arguments{
\item{state}{state geo_id as a string of two numbers}

\item{year}{vintage year of requested file as a string of four numbers}
}
\value{
path to local cache location of the downlaoded unzipped files
}
\description{
This function is fragile, expects very specific arguments.
}
