% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-census-geocoder.R
\name{census_geocoder_api_benchmarks}
\alias{census_geocoder_api_benchmarks}
\title{Retrieve census geocoder benchmarks}
\usage{
census_geocoder_api_benchmarks(endpoint = census_geocoder_api_endpoint())
}
\arguments{
\item{endpoint}{base url for the service}
}
\value{
list with all benchmarks descriptors
}
\description{
Benchmarks are the version of the data used to geocode
}
