% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-census-geocoder.R
\name{census_geocoder_api_call}
\alias{census_geocoder_api_call}
\title{Function for doing a single API call for a single target}
\usage{
census_geocoder_api_call(
  target,
  endpoint = census_geocoder_api_endpoint(),
  returntype = "locations",
  benchmark = purrr::flatten(purrr::keep(census_geocoder_api_benchmarks(),
    ~.x$isDefault)),
  vintage = purrr::flatten(purrr::keep(census_geocoder_api_vintage()$vintages,
    ~.x$isDefault)),
  cache_dir = workflow::build_dir("census-geocoder-cache")
)
}
\arguments{
\item{target}{list with components for a valid call.}

\item{endpoint}{service endpoint}

\item{returntype}{either "locations" for basic API call or "geographies" if the
call should also return which geographic entities the address exists within.}

\item{benchmark, }{as returned by census_geocoder_api_benchmarks}

\item{vintage, }{as returned by census_geocoder_api_vintage}
}
\value{
a list gently re-formatted from the API call return, including
status, any error messages, and the outcome of geocoding.
}
\description{
Function for doing a single API call for a single target
}
