% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-census-geocoder.R
\name{census_geocoder_api_check_target_components}
\alias{census_geocoder_api_check_target_components}
\title{Internal function: checks that the geocoding target contains the correct components}
\usage{
census_geocoder_api_check_target_components(target)
}
\arguments{
\item{target}{target to check}
}
\value{
string indicating searchtype to use with the API
}
\description{
Checks target components and derives the search type
}
