% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-census-geocoder.R
\name{census_geocoder_api_endpoint}
\alias{census_geocoder_api_endpoint}
\title{Endpoint for the geocoder}
\usage{
census_geocoder_api_endpoint(
  default = "https://geocoding.geo.census.gov/geocoder"
)
}
\arguments{
\item{default}{the actual endpoint, should someone want to change it (?)}
}
\value{
endpoint url
}
\description{
Returns the census geocoder endpoint
}
