% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-census-geocoder.R
\name{census_geocoder_api_vintage}
\alias{census_geocoder_api_vintage}
\title{Retrieve version ("vintage") for a given benchmark}
\usage{
census_geocoder_api_vintage(
  endpoint = census_geocoder_api_endpoint(),
  benchmark = purrr::flatten(purrr::keep(census_geocoder_api_benchmarks(),
    ~.x$isDefault))
)
}
\arguments{
\item{endpoint}{base url for the service}

\item{benchmark}{the benchmark whose vintage we want}
}
\value{
list with all vintage descriptors
}
\description{
Versions within versions...
}
