% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-census-geocoder.R
\name{census_geocoder_batch}
\alias{census_geocoder_batch}
\title{Function to geocode a batch (data frame) of addresses in street/city/state/zip format}
\usage{
census_geocoder_batch(
  batch,
  street = street,
  city = city,
  state = state,
  zip = zip,
  endpoint = census_geocoder_api_endpoint(),
  returntype = "locations",
  benchmark = purrr::flatten(purrr::keep(census_geocoder_api_benchmarks(),
    ~.x$isDefault)),
  vintage = purrr::flatten(purrr::keep(census_geocoder_api_vintage()$vintages,
    ~.x$isDefault)),
  cache_dir = workflow::build_dir("census-geocoder-cache")
)
}
\arguments{
\item{batch}{data frame containing columns with parts of the address}

\item{street}{bare symbol for street column}

\item{city}{bare symbol for city column}

\item{state}{bare symbol for state column}

\item{zip}{bare symbol for zip code column}

\item{cache_file}{optional path to file to use to cache results to avoid submitting them
multiple times. Can be shared between calls or it will be based on a hash of the
batch input data.}
}
\value{
list of geocoded components and resulting tables with some simplified output
}
\description{
Function to geocode a batch (data frame) of addresses in street/city/state/zip format
}
