% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-census-geocoder.R
\name{census_geocoder_flatten_result}
\alias{census_geocoder_flatten_result}
\title{A function to flatten the heavily nested API output slightly}
\usage{
census_geocoder_flatten_result(x)
}
\arguments{
\item{x}{output from census_geocoder_api_call function}
}
\value{
flatter list with a tibble containing the results
}
\description{
An opinionated flattening of the API output
}
