% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-geocoder.R
\name{census_local_tract_geocoder}
\alias{census_local_tract_geocoder}
\title{Geocode a local dataset on the given columns and reference year for spatial data}
\usage{
census_local_tract_geocoder(
  data,
  street = street,
  city = city,
  state = state,
  county = county,
  zip = zip,
  reference_year = "2020"
)
}
\arguments{
\item{data}{data frame to pull address columns from}

\item{street}{unquoted address column name for street address component}

\item{city}{unquoted city column component of the address}

\item{state}{unquoted state column component of the address}

\item{zip}{unquoted zip code column component of the address}

\item{year}{four-digit year string for spatial data reference year}
}
\value{
data frame with the original colums as well as the geocoded location tract
}
\description{
Geocode a local dataset on the given columns and reference year for spatial data
}
