% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{ClearArguments}
\alias{ClearArguments}
\title{Removes all arguments from the job}
\usage{
ClearArguments(job)
}
\arguments{
\item{job}{Job object}
}
\value{
Job with no arguments
}
\description{
Removes all arguments from the job
}
\seealso{
\code{\link{Job}}
}
