% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{ClearDependencies}
\alias{ClearDependencies}
\title{Remove all dependencies}
\usage{
ClearDependencies(workflow)
}
\arguments{
\item{workflow}{The Workflow object}
}
\value{
The Workflow object with no dependencies
}
\description{
Remove all dependencies
}
\seealso{
\code{\link{Workflow}}, \code{\link{Dependency}},
 \code{\link{Depends}}, \code{\link{AddDependency}},
 \code{\link{RemoveDependency}}
}
