% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{ClearExecutables}
\alias{ClearExecutables}
\title{Remove all executables}
\usage{
ClearExecutables(workflow)
}
\arguments{
\item{workflow}{Workflow object}
}
\value{
The Workflow object with no executables
}
\description{
Remove all executables
}
\seealso{
\code{\link{Workflow}}, \code{\link{Executable}},
 \code{\link{AddExecutable}}, \code{\link{RemoveExecutable}}
}
