% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{ClearFiles}
\alias{ClearFiles}
\title{Remove all files}
\usage{
ClearFiles(workflow)
}
\arguments{
\item{workflow}{Workflow object}
}
\value{
The Workflow object with no files
}
\description{
Remove all files
}
\seealso{
\code{\link{Workflow}}, \code{\link{File}}, \code{\link{AddFile}}, \code{\link{RemoveFile}}
}
