% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{ClearJobs}
\alias{ClearJobs}
\title{Remove all jobs}
\usage{
ClearJobs(workflow)
}
\arguments{
\item{workflow}{Workflow object}
}
\value{
The Workflow object with no jobs
}
\description{
Remove all jobs
}
\seealso{
\code{\link{Workflow}}, \code{\link{RemoveJob}}, \code{\link{AddJob}}
}
