% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{ClearTransformations}
\alias{ClearTransformations}
\title{Remove all transformations}
\usage{
ClearTransformations(workflow)
}
\arguments{
\item{workflow}{Workflow object}
}
\value{
The Workflow object with no transformations
}
\description{
Remove all transformations
}
\seealso{
\code{\link{Workflow}}, \code{\link{Transformation}},
 \code{\link{AddTransformation}}, \code{\link{RemoveTransformation}}
}
