% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-metadata.R
\name{column}
\alias{column}
\title{Create a column definition from arbitrary arguments}
\usage{
column(...)
}
\arguments{
\item{...}{arguments to \code{ColumnMetadata} initializer}
}
\value{
ColumnMetadata object
}
\description{
Create a column definition from arbitrary arguments
}
