% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-metadata-class.R
\name{ColumnMetadata}
\alias{ColumnMetadata}
\title{A class representing metadata about a single column}
\description{
Describes column contents
}
\details{
It's metadata
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ColumnMetadata$new()}}
\item \href{#method-merge}{\code{ColumnMetadata$merge()}}
\item \href{#method-as_text}{\code{ColumnMetadata$as_text()}}
\item \href{#method-as_list}{\code{ColumnMetadata$as_list()}}
\item \href{#method-clone}{\code{ColumnMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMetadata$new(
  name,
  definition = NA,
  standard_name = standardize_name(name),
  format = NA,
  units = NA,
  values = NA,
  original_name = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merge"></a>}}
\if{latex}{\out{\hypertarget{method-merge}{}}}
\subsection{Method \code{merge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMetadata$merge(other, priority = "self")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_text"></a>}}
\if{latex}{\out{\hypertarget{method-as_text}{}}}
\subsection{Method \code{as_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMetadata$as_text()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMetadata$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
