% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-metadata.R
\name{combine_definitions}
\alias{combine_definitions}
\title{Combine multiple definitions for the same column}
\usage{
combine_definitions(x)
}
\arguments{
\item{x}{list of definitions}
}
\value{
list of definitions with one entry for each current name
}
\description{
Combine multiple definitions for the same column
}
