% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_execute_test_xml.R
\name{create_execute_test_xml}
\alias{create_execute_test_xml}
\title{Create a PEcAn XML file and use it to run a PEcAn workflow}
\usage{
create_execute_test_xml(model_id, met, site_id, start_date, end_date,
  dbfiles_folder, user_id, output_folder = "batch_test_output",
  pecan_path = getwd(), pft = NULL, ensemble_size = 1,
  sensitivity_variable = "NPP", sensitivity = FALSE,
  db_bety_username = "bety", db_bety_password = "bety",
  db_bety_hostname = "localhost", db_bety_driver = "Postgres")
}
\arguments{
\item{model_id}{(numeric) Model ID (from `models` table)}

\item{met}{(character) Name of meteorology input source (e.g. `"CRUNCEP"`)}

\item{site_id}{(numeric) Site ID (from `sites` table)}

\item{start_date}{(character or date) Run start date}

\item{end_date}{(character or date) Run end date}

\item{dbfiles_folder}{(character) Path to `dbfiles` directory}

\item{user_id}{(numeric) User ID to associate with the workflow}

\item{output_folder}{(character) Path to root directory for storing outputs.
Default = `"batch_test_output"`}

\item{pecan_path}{(character) Path to PEcAn source code. Default is current
working directory.}

\item{pft}{(character) Name of PFT to run. If `NULL` (default), use the first
PFT in BETY associated with the model.}

\item{ensemble_size}{(numeric) Number of ensembles to run. Default = 1.}

\item{sensitivity_variable}{(character) Variable for performing sensitivity
analysis. Default = `"NPP"`}

\item{sensitivity}{(logical) Whether or not to perform a sensitivity analysis
(default = `FALSE`)}

\item{db_bety_username}{(character) BETY username for workflow (default = `"bety"`)}

\item{db_bety_password}{(character) BETY password for workflow (default = `"bety"`)}

\item{db_bety_hostname}{(character) BETY hostname for workflow (default = `"localhost"`)}

\item{db_bety_driver}{(character) BETY DBI driver for workflow (default = `"Postgres"`)}
}
\description{
Create a PEcAn XML file and use it to run a PEcAn workflow
}
\author{
Alexey Shiklomanov, Tony Gardella
}
