% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xml.R
\name{create_xml}
\alias{create_xml}
\alias{create_execute_test_xml}
\title{create_xml}
\usage{
create_execute_test_xml(run_list)
}
\arguments{
\item{run_list}{PEcAn table containing specifications of runs}

\item{overwrite.met, overwrite.fia, overwrite.ic}{logical}
}
\description{
Function to create a viable PEcAn xml from a table containing specifications of a run
}
\author{
Tony Gardella
}
