% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-env-vars.R
\name{default_dir_var}
\alias{default_dir_var}
\title{Default environental variable name for a resource name}
\usage{
default_dir_var(name)
}
\arguments{
\item{name}{name of a resource dir}
}
\value{
string of the corresponding environmental variable
}
\description{
Default environental variable name for a resource name
}
