% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{default_pkglist}
\alias{default_pkglist}
\title{Default place to find list of packages}
\usage{
default_pkglist(path = "~/.Rpackages")
}
\arguments{
\item{path}{where to find list}
}
\value{
path to package list
}
\description{
one package name per line
}
