% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{Dependency}
\alias{Dependency}
\title{A dependency between two nodes in the ADAG}
\usage{
Dependency(parent, child, edge.label = NULL)
}
\arguments{
\item{parent}{The parent job/dax/dag or id}

\item{child}{The child job/dax/dag or id}

\item{edge.label}{A label for the edge (optional)}
}
\value{
Dependency object between parent and child
}
\description{
A dependency between two nodes in the ADAG
}
