% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{Depends}
\alias{Depends}
\title{Add a dependency to the workflow}
\usage{
Depends(workflow, child, parent, edge.label = NULL)
}
\arguments{
\item{workflow}{The Workflow object}

\item{child}{The child job/dax/dag or id}

\item{parent}{The parent job/dax/dag or id}

\item{edge.label}{A label for the edge (optional)}
}
\value{
The Workflow object with the dependency appended
}
\description{
Add a dependency to the workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Dependency}}, \code{\link{AddDependency}}
}
