% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{detect_load_problems}
\alias{detect_load_problems}
\title{Check if packages can be loaded report problems}
\usage{
detect_load_problems(pkgs)
}
\arguments{
\item{pkgs}{vector of package names}
}
\value{
vector of package names for packages with load problems
}
\description{
See \code{workflow::try_load_elsewhere} for details on check.
}
