% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environ-file.R
\name{e_string}
\alias{e_string}
\title{Format name/value pair for .Renviron file}
\usage{
e_string(name, var)
}
\arguments{
\item{name}{.Renviron variable name}

\item{var}{.Renviron variable value}
}
\value{
formatted string
}
\description{
Format name/value pair for .Renviron file
}
