% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{Element}
\alias{Element}
\title{Representation of an XML element for formatting output}
\usage{
Element(name, attrs = list())
}
\arguments{
\item{name}{element name}

\item{attrs}{list of element attributes}
}
\value{
an element object
}
\description{
Representation of an XML element for formatting output
}
