% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environ.R
\name{eval_env}
\alias{eval_env}
\title{Take a dotlist of expressions and evaluate them}
\usage{
eval_env(..., .env = rlang::new_environment(parent = rlang::caller_env()))
}
\arguments{
\item{...}{dotlist of expressions}

\item{.env}{evaluation environment}
}
\value{
list of values created in the enviornment
}
\description{
Evaluation environment can be set, it's a new child of the calling
environment by default.
}
