% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch files from under a host and place files under a set of paths}
\usage{
fetch(urls, paths)
}
\arguments{
\item{urls}{urls on host to fetch from}

\item{paths}{paths to place files in}
}
\value{

}
\description{
Fetch files from under a host and place files under a set of paths
}
