% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-fetch.R
\name{fetch_gadm}
\alias{fetch_gadm}
\title{Fetch GADM files}
\usage{
fetch_gadm(
  code,
  version = gadm_version(),
  host = gadm_url(),
  data_dir = data_dir()
)
}
\arguments{
\item{code}{which country code to fetch for}

\item{version}{version to fetch (defaults)}

\item{host}{host to fetch from (defaults)}

\item{data_dir}{directory to place files under}
}
\value{
list of downloaded files written successfully
}
\description{
Fetch GADM files
}
