% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-resource-tools.R
\name{fetch_uscb}
\alias{fetch_uscb}
\title{Fetch USCB block files}
\usage{
fetch_uscb(
  code,
  version = uscb_version(),
  host = uscb_url(),
  data_dir = data_dir()
)
}
\arguments{
\item{version}{version to fetch (defaults)}

\item{host}{host to fetch from (defaults)}

\item{data_dir}{directory to place files under}
}
\value{
list of downloaded files written successfully
}
\description{
Fetch USCB block files
}
