% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-fetch.R
\name{fetch_uscb_block_gdb}
\alias{fetch_uscb_block_gdb}
\title{Fetch USCB block files}
\usage{
fetch_uscb_block_gdb(year, host = uscb_api_endpoint(), data_dir = data_dir())
}
\arguments{
\item{host}{host to fetch from (defaults)}

\item{data_dir}{directory to place files under}

\item{version}{version to fetch (defaults)}
}
\value{
list of downloaded files written successfully
}
\description{
Fetch USCB block files
}
