% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-fetch.R
\name{fetch_uscb_tiger_state_gdb}
\alias{fetch_uscb_tiger_state_gdb}
\title{Fetch USCB state spatial database}
\usage{
fetch_uscb_tiger_state_gdb(
  year,
  state,
  type = "",
  format = "gdb",
  host = uscb_api_endpoint(),
  data_dir = workflow::data_dir()
)
}
\arguments{
\item{year}{year version to fetch}

\item{state}{two-letter code for state abbreviation to cover}

\item{type}{either empty string (default) or 'edges' to select between the two availble.}

\item{host}{host to fetch from (defaults)}

\item{data_dir}{directory to place files under}
}
\value{
list of downloaded files written successfully
}
\description{
Fetch USCB state spatial database
}
