% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-fetch.R
\name{fetch_uscb_tiger_us_gdb}
\alias{fetch_uscb_tiger_us_gdb}
\title{Fetch USCB US block files}
\usage{
fetch_uscb_tiger_us_gdb(
  year,
  type = "block",
  format = "gdb",
  host = uscb_api_endpoint(),
  data_dir = workflow::data_dir()
)
}
\arguments{
\item{year}{year version to fetch}

\item{type}{either empty string (default) or 'edges' to select between the two availble.}

\item{format}{spatial data format, 'gdb' by default}

\item{host}{host to fetch from (defaults)}

\item{data_dir}{directory to place files under}
}
\value{
list of downloaded files written successfully
}
\description{
Fetch USCB US block files
}
