% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{File}
\alias{File}
\title{A file entry for the DAX-level replica catalog, or a reference to a logical file used by the workflow}
\usage{
File(name)
}
\arguments{
\item{name}{File name}
}
\value{
A \code{File} object
}
\description{
All arguments specify the workflow-level behavior of this File. Job-level
behavior can be defined when adding the File to a Job's uses. If the
properties are not overridden at the job-level, then the workflow-level
values are used as defaults.

If this LFN is to be used as a job's stdin/stdout/stderr then the value
of link is ignored when generating the <std*> tags.
}
\seealso{
\code{\link{AddFile}}, \code{\link{RemoveFile}}
}
