% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{find_files}
\alias{find_files}
\title{Search for files under a root directory}
\usage{
find_files(path, name, extension)
}
\arguments{
\item{path}{root directory to search under}

\item{name}{pattern to search for}

\item{extension}{extension to search for}
}
\value{
full paths to all files
}
\description{
Search for files under a root directory
}
