% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-gadm.R
\name{gadm_gpkg_file_path}
\alias{gadm_gpkg_file_path}
\title{GADM zip file full file path}
\usage{
gadm_gpkg_file_path(
  code = "USA",
  version = gadm_version(),
  data_dir = data_dir()
)
}
\arguments{
\item{code}{country code for the file}

\item{version}{version of GADM for the file}
}
\value{
zip file full path
}
\description{
GADM zip file full file path
}
