% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-gadm.R
\name{gadm_gpkg_file_url}
\alias{gadm_gpkg_file_url}
\title{GADM zip file full URL}
\usage{
gadm_gpkg_file_url(code = "USA", version = gadm_version(), host = gadm_url())
}
\arguments{
\item{code}{country code for the file}

\item{version}{version of GADM for the file}
}
\value{
zip file URL
}
\description{
GADM zip file full URL
}
