% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-resource-locations.R
\name{gadm_gpkg_path}
\alias{gadm_gpkg_path}
\title{Path to gpkg files on GADM host}
\usage{
gadm_gpkg_path(version = gadm_version())
}
\arguments{
\item{version}{version of GADM used in relative path}
}
\value{
relative path to GADM files
}
\description{
Path to gpkg files on GADM host
}
