% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-gadm.R
\name{gadm_gpkg_relative_path}
\alias{gadm_gpkg_relative_path}
\title{GADM zip file relative path}
\usage{
gadm_gpkg_relative_path(code = "USA", version = gadm_version())
}
\arguments{
\item{code}{country code for the file}

\item{version}{version of GADM for the file}
}
\value{
zip file name
}
\description{
GADM zip file relative path
}
