% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-gadm.R
\name{gadm_gpkg_zip}
\alias{gadm_gpkg_zip}
\title{GADM zip file name}
\usage{
gadm_gpkg_zip(code = "USA", version = gadm_version(type = "concatenate"))
}
\arguments{
\item{code}{country code for the file}

\item{version}{version of GADM for the file}
}
\value{
zip file name
}
\description{
GADM zip file name
}
