% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-gadm.R
\name{gadm_url}
\alias{gadm_url}
\title{Host URL for GADM}
\usage{
gadm_url(url = "https://biogeo.ucdavis.edu")
}
\arguments{
\item{url}{alternative url to use, defaults to know location}
}
\value{
url as string
}
\description{
Host URL for GADM
}
