% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-generation.R
\name{generate_document}
\alias{generate_document}
\title{Render an rmarkdown doc (.Rmd)}
\usage{
generate_document(path, ..., type = "html")
}
\arguments{
\item{path}{(relative to project root) to the .Rmd file}

\item{...}{parameters to pass to the markdown document}

\item{type}{file extension on the output (.html)}
}
\value{
path to output document
}
\description{
FIXME: currently type \emph{must} be html... :/
}
\details{
Convention is that project contains the doc itself, garbage ends up in
build_dir and output ends up in artifact dir.
}
