% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datapath.R
\name{get_datapath}
\alias{get_datapath}
\title{get_datapath}
\usage{
get_datapath(envvar_datapath = "DATAPATH")
}
\arguments{
\item{envvar_datapath}{(chr) The name of the environmental variable
in which the path to the data folder is stored}
}
\value{
(chr) The data path stored in the \code{envvar_datapath}
environmental variable
}
\description{
Retrieve the path of the data folder stored in an environmental
variable.
}
\examples{
get_datapath()
}
