% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environ.R
\name{get_default_environ}
\alias{get_default_environ}
\title{Default environment values for the workflow}
\usage{
get_default_environ(
  project_name,
  project_version = "00.00.001",
  project_dir = getwd(),
  build_dir = rappdirs::user_cache_dir(appname = project_name, version =
    project_version),
  artifact_dir = rappdirs::user_data_dir(appname = project_name, version =
    project_version),
  data_dir = rappdirs::user_data_dir(appname = project_name, version = project_version),
  config_dir = rappdirs::user_config_dir(appname = project_name, version =
    project_version),
  r_libs = file.path(build_dir, ".R/library"),
  r_libs_user = file.path(build_dir, ".R/library"),
  r_cran = "https://cloud.r-project.org",
  r_pkg_list = ".Rpackages"
)
}
\arguments{
\item{project_name}{arbitrary project name}

\item{project_version}{three-segment current version}

\item{project_dir}{project source files directory}

\item{build_dir}{project scratch space}

\item{artifact_dir}{specified project outputs will go here}

\item{data_dir}{root directory for reaching project data sets, primarily
used as read-only}

\item{config_dir}{configuration files and tokens stored here}

\item{r_libs}{where library files are found}

\item{r_libs_user}{where library files are found, same as r_libs}

\item{r_cran}{repos to use}

\item{r_pkg_list}{name of file storing R packages list for project}
}
\value{
list of these defaults, modified by args if specified.
}
\description{
Default environment values for the workflow
}
