% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{get_default_packages}
\alias{get_default_packages}
\title{Retrieve list of R's default packages}
\usage{
get_default_packages()
}
\value{
R's list of default packages
}
\description{
Just avoiding brain freeze
}
