% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-structure.R
\name{get_dir}
\alias{get_dir}
\title{Retrieve a URI or directory relative to a named resource}
\usage{
get_dir(..., name = NULL, strict = TRUE)
}
\arguments{
\item{...}{relative URI/path components}

\item{name}{required (specified first and named) name or resource}

\item{strict}{iff FALSE (TRUE by default), failure to find the named resource downgrades to
a temporary filesystem location.}
}
\description{
FIXME: currently relies on \code{fs::path} so \emph{only} works with directories
}
