% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-env-vars.R
\name{get_env}
\alias{get_env}
\title{Retrieve system environment variable}
\usage{
get_env(name)
}
\arguments{
\item{name}{name of variable to retrieve}
}
\value{
variable value to retrieve
}
\description{
Retrieve system environment variable
}
