% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{get_pkglist}
\alias{get_pkglist}
\title{Retrieve the list of packages that should be installed}
\usage{
get_pkglist(path = default_pkglist())
}
\arguments{
\item{path}{path to file with default list}
}
\value{
vector of package names to install
}
\description{
Retrieve the list of packages that should be installed
}
