% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-structure.R
\name{get_resource}
\alias{get_resource}
\title{Retrieve (weak-reference style), the named resource}
\usage{
get_resource(name, path = NULL)
}
\arguments{
\item{name}{name of the resource}

\item{path, }{optionally the path to place the resource at}
}
\value{
path to resource, if available
}
\description{
Retrieve (weak-reference style), the named resource
}
