% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{GetJob}
\alias{GetJob}
\title{Get a Job/DAG/Workflow}
\usage{
GetJob(workflow, jobid)
}
\arguments{
\item{workflow}{Workflow object}

\item{jobid}{Job identification}
}
\value{
Job/DAG/Workflow object
}
\description{
Get a Job/DAG/Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{HasJob}}
}
