% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{HasDependency}
\alias{HasDependency}
\title{Check to see if dependency exists}
\usage{
HasDependency(workflow, dep)
}
\arguments{
\item{workflow}{The Workflow object}

\item{dep}{The dependency object}
}
\value{
If the Workflow contains the dependency
}
\description{
Check to see if dependency exists
}
\seealso{
\code{\link{Workflow}}, \code{\link{Dependency}}
}
