% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{HasExecutable}
\alias{HasExecutable}
\title{Check if executable is in this Workflow}
\usage{
HasExecutable(workflow, executable)
}
\arguments{
\item{workflow}{Workflow object}

\item{executable}{Executable object}
}
\value{
If the executable is in the Workflow
}
\description{
Check if executable is in this Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Executable}}
}
