% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{HasFile}
\alias{HasFile}
\title{Check to see if file is in the Workflow}
\usage{
HasFile(workflow, file)
}
\arguments{
\item{workflow}{Workflow object}

\item{file}{File object}
}
\value{
If the Workflow object contains the file
}
\description{
Check to see if file is in the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{File}}
}
