% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{HasJob}
\alias{HasJob}
\title{Test to see if job is in this Workflow}
\usage{
HasJob(workflow, job)
}
\arguments{
\item{workflow}{Workflow object}

\item{job}{Job/DAG/Workflow object}
}
\value{
If the Job/DAG/Workflow is in the Workflow
}
\description{
The job parameter can be an object or a job ID.
}
\seealso{
\code{\link{Workflow}}, \code{\link{GetJob}}
}
