% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{HasTransformation}
\alias{HasTransformation}
\title{Check to see if transformation is in the Workflow}
\usage{
HasTransformation(workflow, transformation)
}
\arguments{
\item{workflow}{Workflow object}

\item{transformation}{Transformation object}
}
\value{
If the Workflow has the transformation
}
\description{
Check to see if transformation is in the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Transformation}}
}
