% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{init}
\alias{init}
\title{Initiliaze a project workflow}
\usage{
init(dir = getwd(), overwrite = FALSE)
}
\arguments{
\item{dir}{Folder to place template files. Defaults to working directory.}

\item{overwrite}{If TRUE, will overwrite existing files. Use with caution.}
}
\description{
This populates the project with template folders and files.
}
\examples{
\dontrun{
init("tmp")
init("tmp", overwrite = TRUE)
}
}
