% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_result.R
\name{insert_preprocess_result}
\alias{insert_preprocess_result}
\title{Insert Preprocess Result into Settings}
\usage{
insert_preprocess_result(settings, result, input.type)
}
\arguments{
\item{settings}{List. PEcAn settings object.}

\item{result}{preprocess_result or list of preprocess_result.}

\item{input.type}{Character. Input type identifier, e.g., "met", "soil".}
}
\value{
Updated settings object.
}
\description{
Takes a preprocess_result (or list of them) and patches it into the
appropriate location in a PEcAn settings object.
}
