% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{install_packages}
\alias{install_packages}
\title{Install a list of packages}
\usage{
install_packages(
  packages = get_pkglist(),
  lib = assure_user_lib()[1],
  repos = assure_repos()
)
}
\arguments{
\item{packages}{vector of package names to install}
}
\value{
list with namems of installs attempted, successes, and failures
}
\description{
Install a list of packages
}
