% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{Invoke}
\alias{Invoke}
\title{Invoke executable \code{what} when job reaches status \code{when}}
\usage{
Invoke(when, what)
}
\arguments{
\item{when}{Job status}

\item{what}{Executable to be invoked when job reach status \code{when}}
}
\value{
Invoke object
}
\description{
Invoke executable \code{what} when job reaches status \code{when}
}
\details{
The value of \code{what} should be a command that can be executed on the submit host.
The list of valid values for 'when' is:

\tabular{ll}{
 WHEN       \tab  MEANING \cr
==========  \tab  ======================================================= \cr
never       \tab  never invoke \cr
start       \tab  invoke just before job gets submitted. \cr
on_error    \tab  invoke after job finishes with failure (exitcode != 0). \cr
on_success  \tab  invoke after job finishes with success (exitcode == 0). \cr
at_end      \tab  invoke after job finishes, regardless of exit status. \cr
all         \tab  like start and at_end combined. \cr
}
}
\examples{
invoke_1 <- Invoke(Pegasus.When$AT_END, '/usr/bin/mail -s "job done" rafsilva@isi.edu')
invoke_2 <- Invoke(Pegasus.When$ON_ERROR, '/usr/bin/update_db -failure')

}
