% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke_mixin.R
\name{InvokeExecutable}
\alias{InvokeExecutable}
\title{Invoke executable \code{what} when job reaches status \code{when}.}
\usage{
InvokeExecutable(invoke.mixin, when, what)
}
\arguments{
\item{invoke.mixin}{InvokeMixin object}

\item{when}{job status}

\item{what}{executable to be invoked when job reach status \code{when}}
}
\value{
InvokeMixin object with invocation appended to the list of invocations
}
\description{
Invoke executable \code{what} when job reaches status \code{when}.
}
\details{
The value of \code{what} should be a command that can be executed on the submit host.
The list of valid values for 'when' is:

\tabular{ll}{
 WHEN       \tab  MEANING \cr
==========  \tab  ======================================================= \cr
never       \tab  never invoke \cr
start       \tab  invoke just before job gets submitted. \cr
on_error    \tab  invoke after job finishes with failure (exitcode != 0). \cr
on_success  \tab  invoke after job finishes with success (exitcode == 0). \cr
at_end      \tab  invoke after job finishes, regardless of exit status. \cr
all         \tab  like start and at_end combined. \cr
}
}
\seealso{
\code{\link{InvokeMixin}}
}
