% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-gadm.R
\name{is_gadm_file}
\alias{is_gadm_file}
\title{Check if a file matches the known GADM file name format}
\usage{
is_gadm_file(path, type = "gpkg", zipped = TRUE)
}
\arguments{
\item{path}{path to file to check}

\item{type}{file format of GADM file}

\item{zipped}{TRUE iff the file is zipped}
}
\value{
TRUE iff the file matches specifications
}
\description{
Check if a file matches the known GADM file name format
}
