% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_usable_path}
\alias{is_usable_path}
\title{Check if a path is usable}
\usage{
is_usable_path(path)
}
\arguments{
\item{path}{path to check}
}
\value{
TRUE iff path is usable
}
\description{
This can be fiddly, follow the same rules:
}
