% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsDefined}
\alias{IsDefined}
\title{Test whether an object is not \code{NULL} and not \code{NA}}
\usage{
IsDefined(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
If the object is not \code{NULL} and not \code{NA}
}
\description{
Test whether an object is not \code{NULL} and not \code{NA}
}
