% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsEqual}
\alias{IsEqual}
\title{Test whether to values are equal}
\usage{
IsEqual(v1, v2)
}
\arguments{
\item{v1}{First value}

\item{v2}{Second value}
}
\value{
If the values are equal
}
\description{
Test whether to values are equal
}
