% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{Job}
\alias{Job}
\title{This class defines the specifics of a job to run in an abstract manner}
\usage{
Job(name, id = NULL, namespace = NULL, version = NULL, node.label = NULL)
}
\arguments{
\item{name}{The transformation name or Transformation object (required)}

\item{id}{A unique identifier for the job (optional)}

\item{namespace}{The namespace of the transformation (optional)}

\item{version}{The transformation version (optional)}

\item{node.label}{The label for this job to use in graphing (optional)}
}
\value{
The job object
}
\description{
All filename references still refer to logical files. All references
transformations also refer to logical transformations, though
physical location hints can be passed through profiles.
}
\details{
The ID for each job should be unique in the DAX. If it is None, then
it will be automatically generated when the job is added to the DAX.

The name, namespace, and version should match what you have in your
transformation catalog. For example, if namespace="foo" name="bar"
and version="1.0", then the transformation catalog should have an
entry for "foo::bar:1.0".

The name argument can be either a string, or a Transformation object. If
it is a Transformation object, then the job will inherit the name, namespace,
and version from the Transformation.
}
\examples{
sleep <- Job(id="ID0001", name="sleep")
jbsim <- Job(id="ID0002", name="jbsim", namespace="cybershake", version="2.1")
merge <- Job("jbsim")

# You can create a Job based on a Transformation:
  mDiff_xform <- Transformation("mDiff", namespace="montage", version="3.0")
  mDiff_job <- Job(mDiff_xform)

# Or an Executable:
  mDiff_exe <- Executable("mDiff", namespace="montage", version="3.0")
  mDiff_job <- Job(mDiff_exe)

# Several arguments can be added at the same time:
  input <- File("i1.txt")
  output <- File("o1.txt")
  mDiff_job <- AddArguments(mDiff_job, list("-i", input, "-o", output))

# Profiles are added similarly:
  mDiff_job <- AddProfile(mDiff_job, Profile(Pegasus.Namespace$ENV, key='PATH', value='/bin'))

# Adding file uses is simple, and you can override global File attributes:
  mDiff_job <- Uses(mDiff_job, input, Pegasus.Link$INPUT)
  mDiff_job <- Uses(mDiff_job, output, Pegasus.Link$OUTPUT, transfer=TRUE, register=TRUE)

}
\seealso{
\code{\link{AddJob}}, \code{\link{Transformation}}, \code{\link{Executable}},
  \code{\link{File}}, \code{\link{Profile}}
}
