% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-metadata.R
\name{make_column_definition}
\alias{make_column_definition}
\title{Create a column definition from a list of arguments}
\usage{
make_column_definition(x)
}
\arguments{
\item{list}{of arguments to \code{ColumnMetadata} initializer}
}
\value{
ColumnMetadata object
}
\description{
Create a column definition from a list of arguments
}
