% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(file = "", path = "", add.terminal = F)
}
\arguments{
\item{file}{Name of the file you wish to read/write}

\item{path}{A element of the DIR list (i.e., DIR$Data)}

\item{add.terminal}{Logical, defaults to F. Should never be changed}
}
\description{
make_filename
}
\author{
Alistair Dunn, Conor Neilson
}
