% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_job_files.R
\name{merge_job_files}
\alias{merge_job_files}
\title{merge_job_files
Merge multiple job.sh files into one larger file.}
\usage{
merge_job_files(settings, jobs_per_file = 10, outdir = NULL)
}
\arguments{
\item{settings}{PEcAn.settings object with host section.}

\item{jobs_per_file}{the number of files you want to merge.}

\item{outdir}{output directory of merged job files.}
}
\value{

}
\description{
merge_job_files
Merge multiple job.sh files into one larger file.
}
\author{
Dongchen Zhang
}
