% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executable.R, R/file.R, R/generic_functions.R,
%   R/job.R, R/transformation.R, R/workflow.R
\name{Metadata.Executable}
\alias{Metadata.Executable}
\alias{Metadata.File}
\alias{Metadata}
\alias{Metadata.Job}
\alias{Metadata.Transformation}
\alias{Metadata.Workflow}
\title{Declarative metadata addition}
\usage{
\method{Metadata}{Executable}(obj, key, value)

\method{Metadata}{File}(obj, key, value)

Metadata(obj, key, value)

\method{Metadata}{Job}(obj, key, value)

\method{Metadata}{Transformation}(obj, key, value)

\method{Metadata}{Workflow}(obj, key, value)
}
\arguments{
\item{obj}{Object to append the metadata}

\item{key}{The metadata key}

\item{value}{The metadata value}
}
\value{
The object containing the metadata
}
\description{
Declarative metadata addition
}
\seealso{
\code{\link{Executable}}

\code{\link{File}}

\code{\link{Metadata}}

\code{\link{Job}}

\code{\link{Transformation}}

\code{\link{Workflow}}
}
