% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_execute_test_xml.R
\name{model_specific_tags}
\alias{model_specific_tags}
\title{model_specific_tags}
\usage{
model_specific_tags(settings, model.info)
}
\arguments{
\item{settings}{pecan xml settings}

\item{model.info}{model info extracted from bety}
}
\value{
updated settings list
}
\description{
model_specific_tags
}
