% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\docType{data}
\name{msg}
\alias{msg}
\title{List of messages returned by the package}
\format{
An object of class \code{list} of length 14.
}
\usage{
msg
}
\description{
In the functions they are referred to as objects.
}
\keyword{datasets}
