% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{NextJobID}
\alias{NextJobID}
\title{Get an autogenerated ID for the next job}
\usage{
NextJobID(workflow)
}
\arguments{
\item{workflow}{Workflow object}
}
\value{
Abstract workflow object with updated sequence number and the next.id in list format: \code{list(Workflow, next.id)}
}
\description{
Get an autogenerated ID for the next job
}
\seealso{
\code{\link{Workflow}}
}
