% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{path_is_good}
\alias{path_is_good}
\title{Check if a path is usable}
\usage{
path_is_good(path)
}
\arguments{
\item{path}{path to check}
}
\value{
TRUE iff path is usable
}
\description{
This can be fiddly, follow the same rules:
}
