% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Pegasus.Arch}
\alias{Pegasus.Arch}
\title{Architecture types}
\format{
An object of class \code{list} of length 8.
}
\usage{
Pegasus.Arch
}
\description{
Architecture types
}
\seealso{
\code{\link{Executable}}
}
\keyword{datasets}
