% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Pegasus.Link}
\alias{Pegasus.Link}
\title{Linkage attributes}
\format{
An object of class \code{list} of length 5.
}
\usage{
Pegasus.Link
}
\description{
Linkage attributes
}
\seealso{
\code{\link{File}}, \code{\link{Executable}}, \code{\link{Uses}}
}
\keyword{datasets}
