% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Pegasus.Namespace}
\alias{Pegasus.Namespace}
\title{Namespace values recognized by Pegasus}
\format{
An object of class \code{list} of length 8.
}
\usage{
Pegasus.Namespace
}
\description{
Namespace values recognized by Pegasus
}
\seealso{
\code{\link{Executable}}, \code{\link{Transformation}}, \code{\link{Job}}
}
\keyword{datasets}
