% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Pegasus.OS}
\alias{Pegasus.OS}
\title{OS types}
\format{
An object of class \code{list} of length 5.
}
\usage{
Pegasus.OS
}
\description{
OS types
}
\seealso{
\code{\link{Executable}}
}
\keyword{datasets}
