% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Pegasus.Transfer}
\alias{Pegasus.Transfer}
\title{Transfer types for uses}
\format{
An object of class \code{list} of length 3.
}
\usage{
Pegasus.Transfer
}
\description{
Transfer types for uses
}
\seealso{
\code{\link{Executable}}, \code{\link{File}}
}
\keyword{datasets}
