% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Pegasus.When}
\alias{Pegasus.When}
\title{Job states for notifications}
\format{
An object of class \code{list} of length 6.
}
\usage{
Pegasus.When
}
\description{
Job states for notifications
}
\seealso{
\code{\link{Job}}, \code{\link{DAX}}, \code{\link{DAG}}, \code{\link{Invoke}}
}
\keyword{datasets}
