% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_result.R
\name{print.pecan_preprocess_result}
\alias{print.pecan_preprocess_result}
\title{Print a Standard Preprocessing Result}
\usage{
\method{print}{pecan_preprocess_result}(x, ...)
}
\arguments{
\item{x}{A `pecan_preprocess_result` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a Standard Preprocessing Result
}
