% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{Profile}
\alias{Profile}
\title{A Profile captures scheduler-, system-, and environment-specific parameters in a uniform fashion}
\usage{
Profile(namespace, key, value)
}
\arguments{
\item{namespace}{The namespace of the profile}

\item{key}{The key name. Can be anything that responds to as.character()}

\item{value}{The value for the profile. Can be anything that responds to as.character()}
}
\value{
Profile object with the defined key=value pair
}
\description{
A Profile captures scheduler-, system-, and environment-specific
parameters in a uniform fashion. Each profile declaration assigns
a value to a key within a namespace.

Profiles can be added to \code{\link{Job}}, \code{\link{DAX}},
\code{\link{DAG}}, \code{\link{File}}, \code{\link{Executable}}, and \code{\link{PFN}}.
}
\examples{
path <- Profile(Pegasus.Namespace$ENV, 'PATH', '/bin')
vanilla <- Profile(Pegasus.Namespace$CONDOR, 'universe', 'vanilla')
path <- Profile(namespace='env', key='PATH', value='/bin')
path <- Profile('env', 'PATH', '/bin')

}
\seealso{
\code{\link{Pegasus.Namespace}}
}
