% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{RemoveDependency}
\alias{RemoveDependency}
\title{Remove dependency from workflow}
\usage{
RemoveDependency(workflow, dep)
}
\arguments{
\item{workflow}{The Workflow object}

\item{dep}{The dependency object}
}
\value{
The Workflow object without the dependency
}
\description{
Remove dependency from workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Dependency}},
 \code{\link{Depends}}, \code{\link{AddDependency}}
}
