% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{RemoveExecutable}
\alias{RemoveExecutable}
\title{Remove executable from the Workflow}
\usage{
RemoveExecutable(workflow, executable)
}
\arguments{
\item{workflow}{Workflow object}

\item{executable}{Executable object}
}
\value{
The Workflow object without the executable
}
\description{
Remove executable from the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Executable}},
 \code{\link{AddExecutable}}, \code{\link{ClearExecutables}}
}
