% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{RemoveFile}
\alias{RemoveFile}
\title{Remove file from this Workflow}
\usage{
RemoveFile(workflow, file)
}
\arguments{
\item{workflow}{Workflow object}

\item{file}{File object}
}
\value{
The Workflow object without the file
}
\description{
Remove file from this Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{File}}, \code{\link{AddFile}}, \code{\link{ClearFiles}}
}
