% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{RemoveJob}
\alias{RemoveJob}
\title{Remove job from the Workflow}
\usage{
RemoveJob(workflow, job)
}
\arguments{
\item{workflow}{Workflow object}

\item{job}{Job/DAG/Workflow object}
}
\value{
The Workflow object without the Job/DAG/Workflow
}
\description{
Remove job from the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{AddJob}}, \code{\link{ClearJobs}}
}
