% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{RemoveTransformation}
\alias{RemoveTransformation}
\title{Remove transformation from the Workflow}
\usage{
RemoveTransformation(workflow, transformation)
}
\arguments{
\item{workflow}{Workflow object}

\item{transformation}{Transformation object}
}
\value{
The Workflow object without the transformation
}
\description{
Remove transformation from the Workflow
}
\seealso{
\code{\link{Workflow}}, \code{\link{Transformation}},
 \code{\link{AddTransformation}}, \code{\link{ClearTransformations}}
}
