% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{resource_dir}
\alias{resource_dir}
\title{Assure that a resource directory for the named resource exists}
\usage{
resource_dir(name, path)
}
\arguments{
\item{name}{name of environmental variable storing resource location}

\item{path}{location to of resource}
}
\value{
path of known-to-exist resource
}
\description{
This is fiddly, follow the same rules
\enumerate{
\item create if needed
\item don't clobber
\item fail if a file is where a directory should be
\item give a clear message of the issues on failure
}
}
