% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_run_in_parallel.R
\name{run_in_parallel}
\alias{run_in_parallel}
\title{Run several R scripts at the same time}
\usage{
run_in_parallel(paths)
}
\arguments{
\item{paths}{(Character) Paths to one or more R scripts.}
}
\value{
A Character string. "Success" for success, and
an error message for failure.
}
\description{
Scripts are run in separate RStudio Background Jobs.
The function only continues if all scripts succeed;
encountering an Error in a script, or cancelling one
of the jobs in the Background Jobs pane, will cause
the function to return an Error.
}
