% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_run_sequentially.R
\name{run_sequentially}
\alias{run_sequentially}
\title{Run several R scripts one after the other}
\usage{
run_sequentially(paths)
}
\arguments{
\item{paths}{(Character) Paths to one or more R scripts.}
}
\value{
A Character string. "Success" for success, and
an error message for failure.
}
\description{
Each script is run as an RStudio Background Job,
with the next job being started only when the previous
jobs report success. Encountering an Error in a script,
or cancelling one of the jobs in the Background Jobs
pane, will cause the function to return an Error.
}
