% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModule.get.trait.data.R
\name{runModule.get.trait.data}
\alias{runModule.get.trait.data}
\title{Get trait data for all PFTs in a settings list}
\usage{
runModule.get.trait.data(settings)
}
\arguments{
\item{settings}{PEcAn configuration list. Must have class `Settings` or `MultiSettings`}
}
\description{
Get trait data for all PFTs in a settings list
}
