% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-env-vars.R
\name{set_env}
\alias{set_env}
\title{Set system environment variable}
\usage{
set_env(name, value)
}
\arguments{
\item{name}{name of variable to set}

\item{value}{value to set the variable to}
}
\value{
value as set
}
\description{
Set system environment variable
}
