% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-environ-file.R
\name{set_environ}
\alias{set_environ}
\title{Write list of values to .Renviron file in correct format}
\usage{
set_environ(
  ...,
  .env = rlang::new_environment(parent = rlang::caller_env()),
  .Renviron = ".Renviron"
)
}
\arguments{
\item{...}{values to modify from defaults}

\item{.env}{environment to evaluate the arguments in,
child of caller by default}

\item{.Renviron}{path to .Renviron file for the project}
}
\value{
list of environment values, these are also written to the
specified .Renviron file \emph{and} read into the current session
using \code{readRenviron}
}
\description{
Any default-required values missing from inputs are filled
in from defaults (see \code{?get_default_environ})
}
\details{
Warning: the default .Renviron path assumes the function is running
with the working directory set to the project directory.
}
