% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_workflow_started.R
\name{stamp_workflow_started}
\alias{stamp_workflow_started}
\title{Timestamp the start time of a workflow}
\usage{
stamp_workflow_started(workflow_id, con, overwrite = FALSE)
}
\arguments{
\item{workflow_id}{ID of workflow to mark as started}

\item{con}{Database connection object. See [PEcAn.DB::db.open()].}

\item{overwrite}{Logical. If `FALSE` (default), only timestamp if
there is no existing time stamp. If `TRUE`, ignore the existing
timestamp.}
}
\value{
(Invisibly) `data.frame` of corresponding row from
  `workflows` table
}
\description{
Timestamp the start time of a workflow
}
\author{
Alexey Shiklomanov
}
