% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-metadata.R
\name{standardize_name}
\alias{standardize_name}
\title{Default name standardizer for columns}
\usage{
standardize_name(x)
}
\arguments{
\item{x}{colummn name}
}
\value{
standardized column name
}
\description{
Default name standardizer for columns
}
