% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_model_runs.R
\name{start_model_runs}
\alias{start_model_runs}
\alias{runModule_start_model_runs}
\title{Start selected ecosystem model runs within PEcAn workflow}
\usage{
start_model_runs(
  settings,
  write = TRUE,
  stop.on.error = TRUE,
  check_interval = 10
)

runModule_start_model_runs(settings, stop.on.error = TRUE, ...)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{write}{(logical) Whether or not to write to the database. Default TRUE.}

\item{stop.on.error}{logical: Kill entire run if any job throws an error?}

\item{check_interval}{Time in seconds to wait between queries of qsub job
status.}

\item{...}{passed on to start_model_runs}
}
\description{
Note: When using a queue system, PEcAn queries each job individually every
`check_interval` seconds. The default interval of ten seconds was chosen for
small sets of jobs in a fast-moving queue, and if used for large runs it can
create excessive load. If you expect the whole run to take hours,
consider checking much less often (every few minutes?) instead.
}
\section{Functions}{
\itemize{
\item \code{runModule_start_model_runs()}: A lightweight wrapper around `start_model_runs` that takes `write` from
`settings` instead of as a separate argument.

}}
\examples{
\dontrun{
  settings <- PEcAn.settings::read.settings("pecan.xml")
  start_model_runs(settings)
}
}
\author{
Shawn Serbin, Rob Kooper, David LeBauer, Alexey Shiklomanov
}
