% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_project.R
\name{start_project}
\alias{start_project}
\title{start_project}
\usage{
start_project()
}
\value{
DIR, a list of directories that can be called by make_filename (see make_filename to learn how to save and access objects in these directories)
}
\description{
Main function to create your folders and set up your directories. This should be the first function run after you have started a new Rstudio project (if using Rstudio), or after setting your working directory (if not using Rstudio)
}
