% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy-matching.R
\name{top_n_matches}
\alias{top_n_matches}
\title{A function to retrieve the top n candidate matches}
\usage{
top_n_matches(x, candidates, n = 5)
}
\arguments{
\item{x}{vector of strings to match against candidates}

\item{candidates}{vector of strings that entries of x may match}

\item{n}{number of matches to keep}
}
\value{
list of plausible matches
}
\description{
A function to retrieve the top n candidate matches
}
