% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/transformation.R
\name{ToXML}
\alias{ToXML}
\alias{ToXML.Transformation}
\title{Get the XML string for the object}
\usage{
ToXML(obj)

\method{ToXML}{Transformation}(obj)
}
\arguments{
\item{obj}{Object to parse as XML}
}
\value{
The XML string for the object
}
\description{
Get the XML string for the object
}
