% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{trim_extension}
\alias{trim_extension}
\title{Trim off last component of file name after a period}
\usage{
trim_extension(path)
}
\arguments{
\item{path}{path to trim extension from}
}
\value{
trimmed path
}
\description{
No I don't want to always remember the stringr call
}
