% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{try_load_elsewhere}
\alias{try_load_elsewhere}
\title{Check if package can be loaded}
\usage{
try_load_elsewhere(pkg)
}
\arguments{
\item{pkg}{name of the package as a string}
}
\value{
TRUE iff package can be loaded
}
\description{
Loading happens in a separate R process (that should find the same
.Renviron, etc... files as the current one so this should be a reproducible
check.  Returns FALSE unless the package can be loaded
}
\details{
Does not affect state of namespace in the current session!
}
