% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-resource-tools.R
\name{url_path}
\alias{url_path}
\title{Like 'paste' but specificall for gluing URL's}
\usage{
url_path(...)
}
\arguments{
\item{...}{strings and values to compose into the URL}
}
\value{
URL
}
\description{
Like 'paste' but specificall for gluing URL's
}
