% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_api_endpoint}
\alias{uscb_api_endpoint}
\title{Host URL for USCB data}
\usage{
uscb_api_endpoint(url = "https://www2.census.gov")
}
\arguments{
\item{url}{alternative url to use, defaults to know location}
}
\value{
url as string
}
\description{
Host URL for USCB data
}
