% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_block_file_url}
\alias{uscb_block_file_url}
\title{USCB block gdb zip file full URL}
\usage{
uscb_block_file_url(year = 2019, host = uscb_api_endpoint(), ...)
}
\value{
zip file URL
}
\description{
USCB block gdb zip file full URL
}
