% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-resource-locations.R
\name{uscb_block_relative_path}
\alias{uscb_block_relative_path}
\title{Relative path to USCB block spatial data}
\usage{
uscb_block_relative_path(year = 2019, ...)
}
\arguments{
\item{year}{in four digits}

\item{area}{area the file covers}

\item{format}{currently only 'gdb' (default)}

\item{location}{currently only 'tiger' (default)}
}
\value{
relative file path name for the resource
}
\description{
Relative path to USCB block spatial data
}
