% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_relative_path}
\alias{uscb_relative_path}
\title{Relative path to USCB data}
\usage{
uscb_relative_path(
  year = 2019,
  area = "us",
  type = "block",
  format = "gdb",
  location = "tiger"
)
}
\arguments{
\item{year}{in four digits}

\item{area}{area the file covers}

\item{type}{spatial data type (block by default)}

\item{format}{currently only 'gdb' (default)}

\item{location}{currently only 'tiger' (default)}
}
\value{
relative file path name for the resource
}
\description{
Relative path to USCB data
}
