% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_tiger_path}
\alias{uscb_tiger_path}
\title{Resource path munging based on characteristics}
\usage{
uscb_tiger_path(year = 2022, format = "gdb")
}
\arguments{
\item{year}{in four digits}

\item{format}{currently only 'gdb'}
}
\value{
relative path to the resource
}
\description{
Resource path munging based on characteristics
}
