% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_tiger_state_file}
\alias{uscb_tiger_state_file}
\title{File name munging from characteristics}
\usage{
uscb_tiger_state_file(year = 2022, state = "MI", type = "", format = "gdb")
}
\arguments{
\item{year}{in four digits}

\item{state}{two-letter code for state to include}

\item{type}{spatial data type (blank or "edges")}

\item{format}{currently only 'gdb' (default)}
}
\value{
file name for the resource
}
\description{
File name munging from characteristics
}
