% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_tiger_state_relative_path}
\alias{uscb_tiger_state_relative_path}
\title{Relative path to USCB state spatial data}
\usage{
uscb_tiger_state_relative_path(
  year = 2019,
  state = "MI",
  type = "",
  format = "gdb"
)
}
\arguments{
\item{year}{in four digits}

\item{state}{two letter code for state to download}

\item{type}{spatial data type ('' is default, alternatively 'edges')}

\item{format}{currently only 'gdb' (default)}
}
\value{
relative file path name for the resource
}
\description{
Relative path to USCB state spatial data
}
