% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_tiger_us_file}
\alias{uscb_tiger_us_file}
\title{File name munging from characteristics}
\usage{
uscb_tiger_us_file(year = 2022, type = "block", format = "gdb")
}
\arguments{
\item{year}{in four digits}

\item{type}{spatial data type (block by default)}

\item{format}{currently only 'gdb' (default)}
}
\value{
file name for the resource
}
\description{
File name munging from characteristics
}
