% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_tiger_us_relative_path}
\alias{uscb_tiger_us_relative_path}
\title{Relative path to USCB data}
\usage{
uscb_tiger_us_relative_path(year = 2019, type = "block", format = "gdb")
}
\arguments{
\item{year}{in four digits}

\item{type}{spatial data type (block by default)}

\item{format}{currently only 'gdb' (default)}
}
\value{
relative file path name for the resource
}
\description{
Relative path to USCB data
}
