% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_zip_file_path}
\alias{uscb_zip_file_path}
\title{USCB zip code gdb zip full local path}
\usage{
uscb_zip_file_path(year = 2019, ..., data_dir = data_dir())
}
\value{
zip file full path
}
\description{
USCB zip code gdb zip full local path
}
