% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_zip_file_url}
\alias{uscb_zip_file_url}
\title{USCB zip code gdb zip full file path}
\usage{
uscb_zip_file_url(year = 2019, ..., data_dir = data_dir())
}
\value{
zip file full url
}
\description{
USCB zip code gdb zip full file path
}
