% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-uscb.R
\name{uscb_zip_relative_path}
\alias{uscb_zip_relative_path}
\title{Relative path to USCB block zip code spatial data}
\usage{
uscb_zip_relative_path(year = 2019, ...)
}
\arguments{
\item{year}{in four digits}

\item{area}{area the file covers}

\item{format}{currently only 'gdb' (default)}

\item{location}{currently only 'tiger' (default)}
}
\value{
relative file path name for the resource
}
\description{
Relative path to USCB block zip code spatial data
}
