% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{WriteYAML}
\alias{WriteYAML}
\title{Write the Workflow as YAML to a stream}
\usage{
WriteYAML(workflow, out)
}
\arguments{
\item{workflow}{The Workflow object}

\item{out}{The stream object (e.g., \code{stdout()}, or a filename)}
}
\description{
Write the Workflow as YAML to a stream
}
\examples{
workflow <- Workflow('diamond')
WriteYAML(workflow, stdout())
WriteYAML(workflow, 'diamond.yml')

}
\seealso{
\code{\link{Workflow}}
}
