% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{WriteYAMLFile}
\alias{WriteYAMLFile}
\title{Write the Workflow to a YAML file}
\usage{
WriteYAMLFile(workflow, filename)
}
\arguments{
\item{workflow}{The Workflow object}

\item{filename}{Name of the file}
}
\description{
Write the Workflow to a YAML file
}
\seealso{
\code{\link{Workflow}}, \code{\link{WriteYAML}}
}
