% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{xlsx_initialize_workbook}
\alias{xlsx_initialize_workbook}
\title{Initialize an Excel Workbook with a README tab}
\usage{
xlsx_initialize_workbook(filename)
}
\arguments{
\item{filename}{path where the Excel workbook will be written}
}
\description{
The README sheet is intended to serve as documentation for the results
stored in the Excel file.
}
\examples{
xlsx_initialize_workbook("tmp.xlsx")
}
\seealso{
Other functions for writing to excel: 
\code{\link{xlsx_write_table}()}
}
\concept{functions for writing to excel}
