% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{xlsx_write_table}
\alias{xlsx_write_table}
\title{Write a data frame to an Excel tab}
\usage{
xlsx_write_table(df, filename, sheet = NULL)
}
\arguments{
\item{df}{data frame to write to the Excel worksheet}

\item{filename}{path where the Excel workbook will be written}

\item{sheet}{name to use for Excel worksheet. If NULL (default) the name of
the df will be used.}
}
\description{
Requires an existing Excel file, preferably created using
\code{\link{xlsx_initialize_workbook}}. The selected tab will be removed
(if it already exists) and a new tab will be written.
}
\examples{
xlsx_initialize_workbook("tmp.xlsx")
moria <- data.frame(a = 1:4, b = c("speak", "friend", "and", "enter"))
xlsx_write_table(moria, "tmp.xlsx")
}
\seealso{
Other functions for writing to excel: 
\code{\link{xlsx_initialize_workbook}()}
}
\concept{functions for writing to excel}
