% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-installation.R
\name{assure_package_installation}
\alias{assure_package_installation}
\title{Maks sure the list of packages is installed}
\usage{
assure_package_installation(path = default_pkglist())
}
\arguments{
\item{path}{path to list of packages (one per line)}
}
\value{

}
\description{
Maks sure the list of packages is installed
}
