% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-class.R
\name{DataTable}
\alias{DataTable}
\title{A class representing a table of data based on a (remote?) file}
\description{
Data...

FIXME: definitions should be viewable in their aggregate effect, not just
returned as an object.  Exported but probably used through helper functions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataTable$new()}}
\item \href{#method-contains}{\code{DataTable$contains()}}
\item \href{#method-correct}{\code{DataTable$correct()}}
\item \href{#method-define}{\code{DataTable$define()}}
\item \href{#method-rename}{\code{DataTable$rename()}}
\item \href{#method-mutate}{\code{DataTable$mutate()}}
\item \href{#method-process_corrections}{\code{DataTable$process_corrections()}}
\item \href{#method-process_definitions}{\code{DataTable$process_definitions()}}
\item \href{#method-attributes}{\code{DataTable$attributes()}}
\item \href{#method-clear_definitions}{\code{DataTable$clear_definitions()}}
\item \href{#method-switch_disk_format}{\code{DataTable$switch_disk_format()}}
\item \href{#method-clone}{\code{DataTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$new(
  uri,
  rpath,
  retrieve = rdrop2::drop_download(path = .source_path, local_path = .local_path),
  load = haven::read_sas(data_file = .local_path),
  .data_dir = workflow::data_dir(),
  .build_dir = workflow::build_dir(),
  .artifact_dir = workflow::artifact_dir(),
  .binary_format = "qs",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-contains"></a>}}
\if{latex}{\out{\hypertarget{method-contains}{}}}
\subsection{Method \code{contains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$contains(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-correct"></a>}}
\if{latex}{\out{\hypertarget{method-correct}{}}}
\subsection{Method \code{correct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$correct(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-define"></a>}}
\if{latex}{\out{\hypertarget{method-define}{}}}
\subsection{Method \code{define()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$define(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rename"></a>}}
\if{latex}{\out{\hypertarget{method-rename}{}}}
\subsection{Method \code{rename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$rename(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-mutate}{}}}
\subsection{Method \code{mutate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$mutate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process_corrections"></a>}}
\if{latex}{\out{\hypertarget{method-process_corrections}{}}}
\subsection{Method \code{process_corrections()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$process_corrections()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process_definitions"></a>}}
\if{latex}{\out{\hypertarget{method-process_definitions}{}}}
\subsection{Method \code{process_definitions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$process_definitions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-attributes"></a>}}
\if{latex}{\out{\hypertarget{method-attributes}{}}}
\subsection{Method \code{attributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$attributes(k, v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_definitions"></a>}}
\if{latex}{\out{\hypertarget{method-clear_definitions}{}}}
\subsection{Method \code{clear_definitions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$clear_definitions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-switch_disk_format"></a>}}
\if{latex}{\out{\hypertarget{method-switch_disk_format}{}}}
\subsection{Method \code{switch_disk_format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$switch_disk_format(to = "qs")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
