% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-resource-locations.R
\name{gadm_version}
\alias{gadm_version}
\title{Current GADM version}
\usage{
gadm_version(major = 3, minor = 6, type = "dot")
}
\arguments{
\item{major}{major version to use}

\item{minor}{minor version to use}

\item{type}{return format, 'dot' means '{major}.{minor}',
'concatenate' means '{major}{minor}', 'major' means '{major}',
and 'minor' means '{minor}'}
}
\value{
gadm version in requested format

@export
}
\description{
Current GADM version
}
