% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(file = "", path = "", add.terminal = F)
}
\arguments{
\item{add.terminal}{}
}
\description{
make_filename
}
\author{
Alistair Dunn, Conor Neilson
}
