% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(file = "", path = "", add.terminal = F)
}
\arguments{
\item{file}{Name of the file you wish to read/write}

\item{path}{A element of the DIR list (i.e., DIR$Data)}

\item{add.terminal}{Logical, defaults to F. Should never be changed}
}
\value{
A complete filename that can be passed with DIR$xxx
}
\description{
Note, make_filename is a useless function in isolation. The usefullness of it comes with being combined with assign_directories and start_project. I recommend reading the vignette to get started.
}
\examples{
\dontrun{
data(iris)
write.csv(iris, make_filename("myiris.csv"), DIR$Data)
}
}
\author{
Alistair Dunn, Conor Neilson
}
