% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-structure.R
\name{resource}
\alias{resource}
\title{Retrieve (weak-reference style), the named resource}
\usage{
resource(name, path = NULL)
}
\arguments{
\item{name}{name of the resource}

\item{path, }{optionally the relative path above the resource}
}
\value{
path to resource, if available
}
\description{
Here a resource is a named location on a filesystem, can be
passed to fs::path, etc..., similar to here::here but doesn't do
a complex search and allows out-of-tree references.
}
