% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{run_html}
\title{Source an R script, storing rendered output as .md in a log subfolder}
\usage{
run(script_path)

run_html(script_path)
}
\arguments{
\item{script_path}{file path of R script to be executed}
}
\description{
Uses \code{\link[rmarkdown]{render}} to create a log of the sourced script.
The markdown output is formatted in a way that works well for github. If not
using github, you may want to use \code{\link{run_html}} instead.
}
\examples{
# workflow::run("path/to/script.R")
}
\concept{functions to run scripts}
