% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModule.run.write.configs.R
\name{runModule.run.write.configs}
\alias{runModule.run.write.configs}
\title{Generate model-specific run configuration files for one or more PEcAn runs}
\usage{
runModule.run.write.configs(settings, overwrite = TRUE)
}
\arguments{
\item{settings}{a PEcAn Settings or MultiSettings object}

\item{overwrite}{logical: Replace config files if they already exist?}
}
\value{
A modified settings object, invisibly
}
\description{
Generate model-specific run configuration files for one or more PEcAn runs
}
