% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_model_runs.R
\name{start_model_runs}
\alias{start_model_runs}
\alias{runModule_start_model_runs}
\title{Start selected ecosystem model runs within PEcAn workflow}
\usage{
start_model_runs(settings, write = TRUE, stop.on.error = TRUE)

runModule_start_model_runs(settings, stop.on.error = TRUE)
}
\arguments{
\item{settings}{pecan settings object}

\item{write}{(logical) Whether or not to write to the database. Default TRUE.}

\item{stop.on.error}{Throw error if _any_ of the runs fails. Default TRUE.}
}
\description{
Start selected ecosystem model runs within PEcAn workflow
}
\section{Functions}{
\itemize{
\item \code{runModule_start_model_runs()}: A lightweight wrapper around `start_model_runs` that takes `write` from
`settings` instead of as a separate argument.

}}
\examples{
\dontrun{
start_model_runs(settings)
}
}
\author{
Shawn Serbin, Rob Kooper, David LeBauer, Alexey Shiklomanov
}
