% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{unzip_file}
\alias{unzip_file}
\title{Unzip a .zip file in the directory hosting the .zip file}
\usage{
unzip_file(path, target)
}
\arguments{
\item{path}{path to file to unzip}
}
\value{
paths of extracted files
}
\description{
Unzip a .zip file in the directory hosting the .zip file
}
