% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-resource-locations.R
\name{uscb_path}
\alias{uscb_path}
\title{Resource path munging based on characteristics}
\usage{
uscb_path(year = 2019, format = "gdb", location = "tiger")
}
\arguments{
\item{year}{in four digits}

\item{format}{currently only 'gdb'}

\item{location}{currently only 'tiger' (default)}
}
\value{
relative path to the resource
}
\description{
Resource path munging based on characteristics
}
